--- spec/support/files/draft_export.csv
@@ -1,5 +1,0 @@
-"UserID","UserNo","Name","Email","IP Address","Unique ID","Started","Ended","clientID","Q11348125","Q11348119","Q11348120","Q11348121","Q11348122","Q11348123","Q11348124"
-"","","","","","","","","","","","","","","",""
-"","","","","","","","","","","","","","","",""
-"100100102","1","","","","","01/05/2020 08:45:00","01/05/2020 09:00:00","","Yes","A question?","Jane Doe","Greater London","jane@example.com","+447123456789","In writing, to be read out at the press conference"
-"100100103","2","","","","","01/05/2020 08:46:00","01/05/2020 09:00:00","","Yes","A question?","Jane Doe","Greater London","jane@example.com","+447123456789","In writing, to be read out at the press conference"
