--- spec/support/files/live_export.csv
@@ -1,4 +1,0 @@
-"UserID","UserNo","Name","Email","IP Address","Unique ID","Started","Ended","clientID","Q11312895","Q11288904","Q11289065","Q11312915","Q11289069","Q11312922","Q11324295"
-"","","","","","","","","","","","","","","",""
-"","","","","","","","","","","","","","","",""
-"100100101","1","","","","","01/05/2020 08:45:00","01/05/2020 09:00:00","","Yes","A question?","Jane Doe","Greater London","jane@example.com","+447123456789","In writing, to be read out at the press conference"
